Nova action do store books
